+++ src/etoolbox/utils/logging_utils.py
@@ -179,7 +179,7 @@ class JSONFormatter(logging.Formatter):
         always_fields = {
             "message": record.getMessage(),
             "timestamp": dt.datetime.fromtimestamp(
-                record.created, tz=dt.timezone.utc
+                record.created, tz=dt.UTC
             ).isoformat(),
         }
         if record.exc_info is not None: